refactor : divide badge apis #48
