+++ Trim-Api/src/main/java/trim/api/domains/badge/controller/BadgeApiController.java
@@ -7,31 +7,23 @@ import org.springframework.web.bind.annotation.*;
 import trim.api.common.dto.ApiResponseDto;
 import trim.api.domains.badge.service.CountUpBadgeOfWritingBoardUseCase;
 import trim.api.domains.badge.service.GetAllBadgesByMemberUseCase;
-import trim.api.domains.badge.service.GetAllBadgesUseCase;
 import trim.api.domains.badge.service.UpgradeBadgeLevelUseCase;
 import trim.api.domains.badge.vo.response.BadgeDetailResponse;
-import trim.api.domains.badge.vo.response.BadgeResponse;
 import trim.common.util.EnumConvertUtil;
 import trim.domains.badge.dao.entity.BadgeContent;
 
 import java.util.List;
 
-@Tag(name = "[뱃지]")
+@Tag(name = "[뱃지🔑]")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/badges")
 public class BadgeApiController {
 
-    private final GetAllBadgesUseCase getAllBadgesUseCase;
     private final UpgradeBadgeLevelUseCase upgradeBadgeLevelUseCase;
     private final CountUpBadgeOfWritingBoardUseCase countUpBadgeOfWritingBoardUseCase;
     private final GetAllBadgesByMemberUseCase getAllBadgesByMemberUseCase;
 
-    @Operation(summary = "모든 뱃지를 조회합니다.")
-    @GetMapping
-    public ApiResponseDto<List<BadgeResponse>> getAllBadges() {
-        return ApiResponseDto.onSuccess(getAllBadgesUseCase.execute());
-    }
 
     @Operation(summary = "모든 뱃지를 조회합니다. 이때 사용자의 미션 상태를 반영하여 값을 가져옵니다.")
     @GetMapping("/members/{memberId}")
@@ -54,6 +46,5 @@ public class BadgeApiController {
         return ApiResponseDto.onSuccess(countUpBadgeOfWritingBoardUseCase
                 .execute(badgeContent, memberId));
     }
-
 
 }
